# year2024/day22: improve runtime of part 2 from 2.6s to 130ms.