After some discussion and looking at other people's code in the #adventofcode
Slack channel, I found out that there is a faster algorithm than I initially
had, so I added it. It improved the runtime from ~2.6s to ~130ms -- an
improvement of 20x!

    goos: darwin
    goarch: arm64
    pkg: go.saser.se/adventofgo/year2024/day22
    cpu: Apple M2 Pro
            │   base.txt   │               new.txt               │
            │    sec/op    │   sec/op     vs base                │
    Part1-12    7.386m ± 1%   7.549m ± 1%   +2.20% (p=0.000 n=10)
    Part2-12   2558.7m ± 1%   126.0m ± 1%  -95.08% (p=0.000 n=10)
    geomean     137.5m        30.84m       -77.57%

            │   base.txt   │                new.txt                 │
            │     B/op     │     B/op      vs base                  │
    Part1-12     24.00 ± 0%     24.00 ± 0%        ~ (p=1.000 n=10) ¹
    Part2-12   119.1Mi ± 0%   165.3Mi ± 0%  +38.79% (p=0.000 n=10)
    geomean    53.47Ki        62.99Ki       +17.81%
    ¹ all samples are equal

            │   base.txt   │                new.txt                │
            │  allocs/op   │  allocs/op   vs base                  │
    Part1-12     2.000 ± 0%    2.000 ± 0%        ~ (p=1.000 n=10) ¹
    Part2-12   156.63k ± 0%   12.79k ± 0%  -91.83% (p=0.000 n=10)
    geomean      559.7         160.0       -71.42%
    ¹ all samples are equal